pyproject.toml
[project]
name = "my-finance-ui"
version = "0.1.0"
description = "Streamlit front end for My Finance"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "plotly>=6.6.0",
    "st-styled>=0.4.0",
    "streamlit>=1.56.0",
]
